# apps/api-go/api/openapi.yaml
openapi: 3.0.0
paths:
...
  /api/v1/courses:
    get:
      operationId: CoursesController_findAll_v1
      parameters:
      - name: sort
        required: false
        in: query
        schema:
          enum:
          - newest
          - recent
          - rating
          - popular
          type: string
      - name: search
        required: false
        in: query
        description: 标题/描述/讲师模糊搜索
        schema: {}
      - name: courseType
        required: false
        in: query
        description: 课程类型过滤
        schema:
          enum:
          - own
          - partner
          - public
          - third_party
          type: string
      - name: status
        required: false
        in: query
        description: 课程状态过滤
        schema:
          enum:
          - draft
          - published
          - archived
          type: string
      responses:
        '200':
          description: ''
      summary: 获取课程列表（公开）
      tags:
      - courses
    post:
      operationId: CoursesController_create_v1
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateCourseDto'
      responses:
        '201':
          description: ''
      security:
      - bearer: []
      summary: 创建课程（管理员）
      tags:
      - courses
  /api/v1/courses/{id}:
    get:
      operationId: CoursesController_findOne_v1
      parameters:
      - name: id
        required: true
        in: path
        description: 课程ID
        schema:
          type: string
      responses:
        '200':
          description: ''
      summary: 获取课程详情（公开；管理员可见草稿）
      tags:
      - courses
    patch:
      operationId: CoursesController_update_v1
      parameters:
      - name: id
        required: true
        in: path
        description: 课程ID
        schema:
          type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateCourseDto'
      responses:
        '200':
          description: ''
      security:
      - bearer: []
      summary: 更新课程（管理员）
      tags:
      - courses
    delete:
      operationId: CoursesController_delete_v1
      parameters:
      - name: id
        required: true
        in: path
        description: 课程ID
        schema:
          type: string
      responses:
        '200':
          description: ''
      security:
      - bearer: []
      summary: 删除课程（管理员）
      tags:
      - courses
  /api/v1/courses/{id}/degrees:
    post:
      operationId: CoursesController_linkDegrees_v1
      parameters:
      - name: id
        required: true
        in: path
        description: 课程ID
        schema:
          type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LinkDegreesDto'
      responses:
        '201':
          description: ''
      security:
      - bearer: []
      summary: 把课程追加到指定学位末尾（管理员）
      tags:
      - courses
...
  /api/v1/degrees/{id}/courses:
    post:
      operationId: DegreesController_linkCourses_v1
      parameters:
      - name: id
        required: true
        in: path
        description: 学位ID
        schema:
          type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LinkCoursesDto'
      responses:
        '201':
          description: ''
      security:
      - bearer: []
      summary: 给学位绑定课程（管理员）
      tags:
      - degrees
...
  /api/v1/practices/courses/{courseId}:
    get:
      operationId: PracticesController_getProjectsByCourse_v1
      parameters:
      - name: courseId
        required: true
        in: path
        description: 课程ID
        schema:
          type: string
      responses:
        '200':
          description: ''
      summary: 获取课程的实践项目列表
      tags:
      - practices
...
  /api/v1/practices/admin/courses/{courseId}:
    get:
      operationId: PracticesController_getAdminProjectsByCourse_v1
      parameters:
      - name: courseId
        required: true
        in: path
        description: 课程ID
        schema:
          type: string
      responses:
        '200':
          description: ''
      security:
      - bearer: []
      summary: 获取课程的全部实践项目（管理员）
      tags:
      - practices
...
  /api/v1/progress/courses/{courseId}:
    get:
      operationId: ProgressController_getCourseProgress_v1
      parameters:
      - name: courseId
        required: true
        in: path
        description: 课程ID
        schema:
          type: string
      responses:
        '200':
          description: ''
      security:
      - bearer: []
      summary: 获取我在某课程的进度
      tags:
      - progress
...
info:
  title: AI Academy API
  description: AI and LLM education platform API
  version: '1.0'
  contact: {}
...
servers: []
components:
  securitySchemes:
    bearer:
      scheme: bearer
      bearerFormat: JWT
      type: http
  schemas:
...
    CreateResourceDto:
      type: object
      properties:
        title:
          type: string
        url:
          type: string
          description: 资源 URL (http/https)
        type:
          type: string
          enum:
          - pdf
          - code
          - link
          - video
          - audio
        isLocked:
          type: boolean
          default: true
      required:
      - title
      - url
      - type
    NestedCourseLessonDto:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
        videoUrl:
          type: string
          description: 课时视频 URL (http/https)
        videoDuration:
          type: number
        isPreview:
          type: boolean
          default: false
        orderIndex:
          type: number
          description: 排序索引
        resources:
          description: 关联资源
          type: array
          items:
            $ref: '#/components/schemas/CreateResourceDto'
      required:
      - title
    CreateChapterDto:
      type: object
      properties:
        title:
          type: string
          description: 章节标题
        description:
          type: string
          description: 章节描述
        orderIndex:
          type: number
          description: 排序索引
        lessons:
          description: 章节下的课时
          type: array
          items:
            $ref: '#/components/schemas/NestedCourseLessonDto'
      required:
      - title
      - orderIndex
    CreateCourseDto:
      type: object
      properties:
        title:
          type: string
          description: 课程标题
        description:
          type: string
          description: 课程描述
        learningPoints:
          type: string
          description: 学习要点
        instructor:
          type: string
          description: 讲师名称
        instructorId:
          type: string
          description: 讲师用户 UUID
        level:
          type: string
          enum:
          - Beginner
          - Intermediate
          - Advanced
          - Expert
          description: 难度等级
        duration:
          type: string
          description: 学习时长（人类可读，例如 "8h"）
        thumbnail:
          type: string
          description: 课程缩略图 URL
        tags:
          type: string
          description: 标签（逗号分隔或 JSON 字符串）
        costType:
          type: string
          enum:
          - free
          - paid
          - charity
          description: 计费类型
        price:
          type: number
          description: 价格（元，0 表示免费）
        status:
          type: string
          enum:
          - draft
          - published
          - archived
          description: 课程状态
        courseType:
          type: string
          enum:
          - own
          - partner
          - public
          - third_party
          description: 课程类型
        externalUrl:
          type: string
          description: 外链课程 URL（外部跳转类型课程用）
          maxLength: 500
        sourceVideoUrl:
          type: string
          description: 原始视频源 URL（导入用）
          maxLength: 500
        sourcePlatform:
          type: string
          description: 来源平台标识
          maxLength: 20
        industryId:
          type: string
          description: 行业 UUID
        categoryId:
          type: string
          description: 课程分类 UUID
        chapters:
          description: 课程章节（含课时和资源）
          type: array
          items:
            $ref: '#/components/schemas/CreateChapterDto'
      required:
      - title
      - description
      - learningPoints
      - instructor
      - level
      - duration
      - thumbnail
      - tags
      - costType
      - price
    UpdateCourseDto:
      type: object
      properties:
        title:
          type: string
          description: 课程标题
        description:
          type: string
          description: 课程描述
        learningPoints:
          type: string
          description: 学习要点
        instructor:
          type: string
          description: 讲师名称
        instructorId:
          type: string
          description: 讲师用户 UUID
        level:
          type: string
          enum:
          - Beginner
          - Intermediate
          - Advanced
          - Expert
          description: 难度等级
        duration:
          type: string
          description: 学习时长（人类可读，例如 "8h"）
        thumbnail:
          type: string
          description: 课程缩略图 URL
        tags:
          type: string
          description: 标签（逗号分隔或 JSON 字符串）
        costType:
          type: string
          enum:
          - free
          - paid
          - charity
          description: 计费类型
        price:
          type: number
          description: 价格（元，0 表示免费）
        status:
          type: string
          enum:
          - draft
          - published
          - archived
          description: 课程状态
        courseType:
          type: string
          enum:
          - own
          - partner
          - public
          - third_party
          description: 课程类型
        externalUrl:
          type: string
          description: 外链课程 URL（外部跳转类型课程用）
          maxLength: 500
        sourceVideoUrl:
          type: string
          description: 原始视频源 URL（导入用）
          maxLength: 500
        sourcePlatform:
          type: string
          description: 来源平台标识
          maxLength: 20
        industryId:
          type: string
          description: 行业 UUID
        categoryId:
          type: string
          description: 课程分类 UUID
        chapters:
          description: 课程章节（含课时和资源）
          type: array
          items:
            $ref: '#/components/schemas/CreateChapterDto'
      required:
      - title
      - description
      - learningPoints
      - instructor
      - level
      - duration
      - thumbnail
      - tags
      - costType
      - price
    LinkDegreesDto:
      type: object
      properties:
        degreeIds:
          description: 要追加到的学位 UUID 列表
          type: array
          items:
            type: string
      required:
      - degreeIds
...
    DegreeCourseLinkDto:
      type: object
      properties:
        courseId:
          type: string
          description: 课程 UUID
        orderIndex:
          type: number
          description: 在学位内的排序索引
      required:
      - courseId
      - orderIndex
    LinkCoursesDto:
      type: object
      properties:
        courses:
          description: 要绑定的课程列表（含排序）
          type: array
          items:
            $ref: '#/components/schemas/DegreeCourseLinkDto'
      required:
      - courses
...